+++ hipcub/include/hipcub/backend/rocprim/iterator/tex_ref_input_iterator.hpp
@@ -1,5 +1,5 @@
 /******************************************************************************
- * Copyright (c) 2010-2011, Duane Merrill.  All rights reserved.
+ * Copyright (c) 2011, Duane Merrill.  All rights reserved.
  * Copyright (c) 2011-2018, NVIDIA CORPORATION.  All rights reserved.
  * Modifications Copyright (c) 2017-2021, Advanced Micro Devices, Inc.  All rights reserved.
  *
@@ -27,20 +27,18 @@
  *
  ******************************************************************************/
 
-#ifndef HIPCUB_ROCPRIM_ITERATOR_TEX_OBJ_INPUT_ITERATOR_HPP_
-#define HIPCUB_ROCPRIM_ITERATOR_TEX_OBJ_INPUT_ITERATOR_HPP_
+#ifndef HIPCUB_ROCPRIM_ITERATOR_TEX_REF_INPUT_ITERATOR_HPP_
+#define HIPCUB_ROCPRIM_ITERATOR_TEX_REF_INPUT_ITERATOR_HPP_
 
 #include <iterator>
 #include <iostream>
 
 #include "../../../config.hpp"
 
-#if (THRUST_VERSION >= 100700)
-    // This iterator is compatible with Thrust API 1.7 and newer
+#if (THRUST_VERSION >= 100700)    // This iterator is compatible with Thrust API 1.7 and newer
     #include <thrust/iterator/iterator_facade.h>
     #include <thrust/iterator/iterator_traits.h>
 #endif // THRUST_VERSION
-
 
 #include <rocprim/iterator/texture_cache_iterator.hpp>
 
@@ -48,9 +46,10 @@ BEGIN_HIPCUB_NAMESPACE
 
 template<
     typename T,
+    int UNIQUE_ID, // Unused parameter for compatibility with original definition in cub
     typename OffsetT = std::ptrdiff_t
 >
-class TexObjInputIterator : public ::rocprim::texture_cache_iterator<T, OffsetT>
+class TexRefInputIterator : public ::rocprim::texture_cache_iterator<T, OffsetT>
 {
     public:
     template<class Qualified>
@@ -68,15 +67,15 @@ class TexObjInputIterator : public ::rocprim::texture_cache_iterator<T, OffsetT>
     }
 
     HIPCUB_HOST_DEVICE inline
-    ~TexObjInputIterator() = default;
+    ~TexRefInputIterator() = default;
 
     HIPCUB_HOST_DEVICE inline
-    TexObjInputIterator() : ::rocprim::texture_cache_iterator<T, OffsetT>()
+    TexRefInputIterator() : ::rocprim::texture_cache_iterator<T, OffsetT>()
     {
     }
 
     HIPCUB_HOST_DEVICE inline
-    TexObjInputIterator(const ::rocprim::texture_cache_iterator<T, OffsetT> other)
+    TexRefInputIterator(const ::rocprim::texture_cache_iterator<T, OffsetT> other)
         : ::rocprim::texture_cache_iterator<T, OffsetT>(other)
     {
     }